requirements.txt
streamlit
openai
langchain
langchain-community
PyPDF2
faiss-cpu
python-dotenv
tiktoken

